vapt commit

+++ app/services/platformsvc/user/pusersvc.js
@@ -1,7 +1,6 @@
 import PpUserSvcDB from "./pusersvc_db.js";
 
 const PLATFORM_ACCOUNT_ID = "ffffffff-ffff-ffff-ffff-ffffffffffff";
-const PLATFORM_FLEET_ID = "ffffffff-ffff-ffff-ffff-ffffffffffff";
 
 export default class PUserSvc {
   constructor(pgPoolI, logger) {
@@ -46,18 +45,32 @@ export default class PUserSvc {
   }
 
   async AddUserPlatformRole(userid, roleids) {
+    const rootFleetId = await this.pUserSvcDB.getRootFleetId(
+      PLATFORM_ACCOUNT_ID
+    );
+    if (!rootFleetId) {
+      throw new Error("Platform account root fleet not found");
+    }
+
     return await this.pUserSvcDB.addUserRole(
       PLATFORM_ACCOUNT_ID,
-      PLATFORM_FLEET_ID,
+      rootFleetId,
       userid,
       roleids
     );
   }
 
   async RemoveUserPlatformRole(userid, roleid) {
+    const rootFleetId = await this.pUserSvcDB.getRootFleetId(
+      PLATFORM_ACCOUNT_ID
+    );
+    if (!rootFleetId) {
+      throw new Error("Platform account root fleet not found");
+    }
+
     return await this.pUserSvcDB.removeUserRole(
       PLATFORM_ACCOUNT_ID,
-      PLATFORM_FLEET_ID,
+      rootFleetId,
       userid,
       roleid
     );
@@ -69,5 +82,57 @@ export default class PUserSvc {
 
   async CheckSuperAdminRole(userid) {
     return await this.pUserSvcDB.checkSuperAdminRole(userid);
+  }
+
+  async ListPendingUsers() {
+    return await this.pUserSvcDB.listPendingUsers();
+  }
+
+  async ListDoneUsers() {
+    return await this.pUserSvcDB.listDoneUsers();
+  }
+  async AddReviewDoneUser(userData) {
+    return await this.pUserSvcDB.addReviewDoneUser(userData);
+  }
+
+
+
+  async AddReviewPendingUser(userData) {
+    return await this.pUserSvcDB.addReviewPendingUser(userData);
+  }
+
+  async AddUserInfo(userid, userinfo, createdby) {
+    return await this.pUserSvcDB.addUserInfo(userid, userinfo, createdby);
+  }
+
+  async GetMetadataOptions() {
+    return await this.pUserSvcDB.getMetadataOptions();
+  }
+
+  async GetUserInfo(userid) {
+    return await this.pUserSvcDB.getUserInfo(userid);
+  }
+
+  async GetPendingUserReviewById(userid) {
+    return await this.pUserSvcDB.getPendingUserReviewById(userid);
+  }
+
+  async DeletePendingUserReviewById(userid) {
+    return await this.pUserSvcDB.deletePendingUserReviewById(userid);
+  }
+  async UpdateReviewPendingUser(userid, updateFields, updatedby) {
+    return await this.pUserSvcDB.updateReviewPendingUser(userid, updateFields, updatedby);
+  }
+
+  async UpdateUserInfo(userid, userinfo, updatedby) {
+    return await this.pUserSvcDB.updateUserInfo(userid, userinfo, updatedby);
+  }
+
+  async checkIsUserAddedToAccount(userid, accountid) {
+    return await this.pUserSvcDB.checkIsUserAddedToAccount(userid, accountid);
+  }
+
+  async checkIsVehicleAddedToAccount(vinno) {
+    return await this.pUserSvcDB.checkIsVehicleAddedToAccount(vinno);
   }
 }
